backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
apscheduler==3.10.4
httpx==0.26.0
aiofiles==23.2.1
cryptography==42.0.2
pyjwt==2.8.0
pymysql==1.1.0
cryptography==42.0.2
oracledb==2.0.1
python-dotenv==1.0.0
jinja2==3.1.3
python-json-logger==2.0.7
